Fixed abstract analytic classes to meet better coding standards my
merging them all together into their own base source and header file. As
a result also fixed some bugs and fixed include issues.

+++ src/core/ace_dataobject.cpp
@@ -300,7 +300,7 @@ const EDataStream& DataObject::stream() const
  * @param size The number of bytes to allocate after this data object's current 
  *             cursor position. 
  */
-void DataObject::allocate(int size)
+void DataObject::allocate(qint64 size)
 {
    EDEBUG_FUNC(this,size)
